# specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Semiont API",
    "version": "0.1.0",
    "description": "Semantic Knowledge Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API Server"
    }
  ],
  "components": {
    "schemas": {
...
      "Motivation": {
        "type": "string",
        "enum": [
          "assessing",
          "bookmarking",
          "classifying",
          "commenting",
          "describing",
          "editing",
          "highlighting",
          "identifying",
          "linking",
          "moderating",
          "questioning",
          "replying",
          "tagging"
        ],
        "description": "W3C Web Annotation motivation vocabulary - https://www.w3.org/TR/annotation-vocab/#motivation"
      },
...
      "CreateAnnotationResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          }
        },
        "required": [
          "annotation"
        ]
      },
      "ListAnnotationsResponse": {
        "type": "object",
        "properties": {
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            }
          },
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "annotations",
          "total",
          "offset",
          "limit"
        ]
      },
...
      "GetAnnotationResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          },
          "resource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolvedResource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "annotation",
          "resource",
          "resolvedResource"
        ]
      },
...
      "CreateAnnotationRequest": {
        "type": "object",
        "properties": {
          "motivation": {
            "$ref": "#/components/schemas/Motivation"
          },
          "target": {
            "type": "object",
            "properties": {
              "source": {
                "type": "string"
              },
              "selector": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TextPositionSelector"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "TextQuoteSelector"
                        ]
                      },
                      "exact": {
                        "type": "string"
                      },
                      "prefix": {
                        "type": "string"
                      },
                      "suffix": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "type",
                      "exact"
                    ]
                  },
                  {
                    "type": "array",
                    "items": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/TextPositionSelector"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": [
                                "TextQuoteSelector"
                              ]
                            },
                            "exact": {
                              "type": "string"
                            },
                            "prefix": {
                              "type": "string"
                            },
                            "suffix": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "type",
                            "exact"
                          ]
                        }
                      ]
                    }
                  }
                ]
              }
            },
            "required": [
              "source",
              "selector"
            ]
          },
          "body": {
            "oneOf": [
              {
                "type": "array",
                "maxItems": 0,
                "description": "Empty array (Phase 1 stub pattern - will be migrated to contain entity tag bodies in Phase 2)"
              },
              {
                "$ref": "#/components/schemas/AnnotationBody",
                "description": "Single body (TextualBody or SpecificResource)"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "minItems": 1,
                "description": "Array of mixed TextualBody (tagging) and SpecificResource (linking) bodies"
              }
            ]
          }
        },
        "required": [
          "target",
          "body",
          "motivation"
        ]
      },
...
      "BodyOperationAdd": {
        "type": "object",
        "properties": {
          "op": {
            "type": "string",
            "const": "add"
          },
          "item": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextualBody"
              },
              {
                "$ref": "#/components/schemas/SpecificResource"
              }
            ]
          }
        },
        "required": [
          "op",
          "item"
        ]
      },
      "BodyOperationRemove": {
        "type": "object",
        "properties": {
          "op": {
            "type": "string",
            "const": "remove"
          },
          "item": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextualBody"
              },
              {
                "$ref": "#/components/schemas/SpecificResource"
              }
            ]
          }
        },
        "required": [
          "op",
          "item"
        ]
      },
      "BodyOperationReplace": {
        "type": "object",
        "properties": {
          "op": {
            "type": "string",
            "const": "replace"
          },
          "oldItem": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextualBody"
              },
              {
                "$ref": "#/components/schemas/SpecificResource"
              }
            ]
          },
          "newItem": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextualBody"
              },
              {
                "$ref": "#/components/schemas/SpecificResource"
              }
            ]
          }
        },
        "required": [
          "op",
          "oldItem",
          "newItem"
        ]
      },
      "UpdateAnnotationBodyRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "description": "Resource ID containing the annotation (required for O(1) Layer 3 lookup)"
          },
          "operations": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BodyOperationAdd"
                },
                {
                  "$ref": "#/components/schemas/BodyOperationRemove"
                },
                {
                  "$ref": "#/components/schemas/BodyOperationReplace"
                }
              ]
            },
            "minItems": 1,
            "description": "Array of body modification operations to apply"
          }
        },
        "required": [
          "resourceId",
          "operations"
        ]
      },
      "UpdateAnnotationBodyResponse": {
        "type": "object",
        "properties": {
          "annotation": {
            "$ref": "#/components/schemas/Annotation"
          }
        },
        "required": [
          "annotation"
        ]
      },
      "DeleteAnnotationRequest": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string",
            "format": "uri",
            "description": "Resource URI containing the annotation (required for O(1) Layer 3 lookup)"
          }
        },
        "required": [
          "resourceId"
        ]
      },
      "TextPositionSelector": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TextPositionSelector"
            ]
          },
          "start": {
            "type": "number",
            "description": "Character offset from resource start"
          },
          "end": {
            "type": "number",
            "description": "Character offset from resource start"
          }
        },
        "required": [
          "type",
          "start",
          "end"
        ]
      },
...
      "ResourceDescriptor": {
        "type": "object",
        "description": "Metadata about a resource (1:1 with its URI). JSON-LD subject is @id. Link to concrete bytes via representations.",
        "required": [
          "@context",
          "@id",
          "name",
          "representations"
        ],
        "additionalProperties": true,
        "properties": {
          "@context": {
            "description": "JSON-LD context; URI, object, or array of these.",
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "string",
                      "format": "uri"
                    },
                    {
                      "type": "object",
                      "additionalProperties": true
                    }
                  ]
                }
              }
            ]
          },
          "@id": {
            "type": "string",
            "format": "uri",
            "description": "Canonical URI/URN of the resource being described."
          },
          "@type": {
            "description": "Type(s) of the resource (IRIs/CURIEs via @context).",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "identifier": {
            "description": "Persistent identifiers (e.g., DOI, URN).",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              },
              {
                "type": "object",
                "properties": {
                  "@id": {
                    "type": "string",
                    "format": "uri"
                  },
                  "value": {
                    "type": "string"
                  },
                  "scheme": {
                    "type": "string"
                  }
                },
                "additionalProperties": true
              }
            ]
          },
          "about": {
            "description": "Topics (IRIs or strings).",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              }
            ]
          },
          "sameAs": {
            "description": "Equivalent/authoritative references.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "isPartOf": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "hasPart": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "license": {
            "type": "string",
            "format": "uri"
          },
          "version": {
            "type": "string"
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "dateModified": {
            "type": "string",
            "format": "date-time"
          },
          "wasDerivedFrom": {
            "description": "W3C PROV - source resources this was derived from",
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            ]
          },
          "wasAttributedTo": {
            "description": "W3C PROV - agents responsible for this resource",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Agent"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            ]
          },
          "conformsTo": {
            "description": "Profile/shape URI this resource description conforms to.",
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            ]
          },
          "availableFormats": {
            "description": "Convenience set summarizing media types across representations.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "representations": {
            "description": "Managed or referenced byte-level renditions of this resource.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Representation"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Representation"
                }
              }
            ]
          },
          "archived": {
            "type": "boolean",
            "description": "Application-specific: Whether this resource is archived"
          },
          "entityTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Application-specific: Entity types for this resource"
          },
          "isDraft": {
            "type": "boolean",
            "description": "Application-specific: Whether this resource is a draft"
          },
          "creationMethod": {
            "type": "string",
            "enum": [
              "api",
              "upload",
              "ui",
              "reference",
              "clone",
              "generated"
            ],
            "description": "Application-specific: How this resource was created"
          },
          "sourceAnnotationId": {
            "type": "string",
            "description": "Application-specific: ID of annotation that triggered generation"
          },
          "sourceResourceId": {
            "type": "string",
            "description": "Application-specific: ID of source resource for clones/derivatives"
          }
        }
      },
...
      "Annotation": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "string",
            "const": "http://www.w3.org/ns/anno.jsonld",
            "default": "http://www.w3.org/ns/anno.jsonld",
            "description": "W3C Web Annotation JSON-LD context"
          },
          "type": {
            "type": "string",
            "const": "Annotation",
            "default": "Annotation",
            "description": "W3C Annotation type"
          },
          "id": {
            "type": "string"
          },
          "motivation": {
            "$ref": "#/components/schemas/Motivation"
          },
          "target": {
            "oneOf": [
              {
                "type": "string",
                "description": "IRI of the entire resource being annotated"
              },
              {
                "$ref": "#/components/schemas/AnnotationTarget"
              }
            ],
            "description": "W3C Web Annotation target - can be a simple IRI string (entire resource) or an object with source and optional selector (fragment)"
          },
          "body": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "maxItems": 0,
                "description": "Empty array for stub references (Phase 1 pattern)"
              },
              {
                "$ref": "#/components/schemas/AnnotationBody",
                "description": "Single body (TextualBody or SpecificResource)"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "description": "Array of mixed TextualBody (tagging) and SpecificResource (linking) bodies"
              }
            ],
            "description": "W3C Web Annotation body - can be empty array, single body, or array of bodies"
          },
          "creator": {
            "$ref": "#/components/schemas/Agent"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          },
          "generator": {
            "$ref": "#/components/schemas/Agent"
          }
        },
        "required": [
          "@context",
          "type",
          "id",
          "motivation",
          "target",
          "body"
        ]
      },
      "AnnotationBody": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/TextualBody"
          },
          {
            "$ref": "#/components/schemas/SpecificResource"
          }
        ],
        "description": "Phase 2: Body can be TextualBody (for entity tags, descriptions) or SpecificResource (for links)"
      },
...
    },
...
  },
  "paths": {
...
    "/annotations/{id}": {
      "get": {
        "summary": "Get Annotation by URI (W3C Content Negotiation)",
        "description": "W3C-compliant globally resolvable annotation URI with content negotiation. Returns JSON-LD for API clients (Accept: application/ld+json or application/json). Redirects to frontend for browsers (Accept: text/html or browser User-Agent). Requires authentication and resourceId query parameter.",
        "tags": [
          "W3C URIs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Annotation ID"
          },
          {
            "name": "resourceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Resource ID or URI containing the annotation"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotation representation (JSON-LD)",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnnotationResponse"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to frontend (for browsers)"
          },
          "400": {
            "description": "Missing resourceId parameter"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
...
    "/api/annotations": {
      "post": {
        "summary": "Create Annotation",
        "description": "Create a new annotation/reference in a resource",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Annotation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnnotationResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Annotations",
        "description": "List all annotations for a resource (requires resourceId for O(1) Layer 3 lookup)",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Resource ID to list annotations for"
            },
            "required": true,
            "description": "Resource ID to list annotations for",
            "name": "resourceId",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Annotations listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAnnotationsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotations/{id}/body": {
      "put": {
        "summary": "Update Annotation Body",
        "description": "Apply fine-grained operations to modify annotation body items (add, remove, or replace TextualBody and SpecificResource items)",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnotationBodyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Annotation body updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAnnotationBodyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotations/{id}": {
      "delete": {
        "summary": "Delete Annotation",
        "description": "Delete an annotation (requires resourceId in body for O(1) Layer 3 lookup)",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Annotation deleted successfully"
          },
          "404": {
            "description": "Annotation not found"
          }
        }
      }
    },
...
  },
...
}